🧪 重构测试Mock架构: 精准线程防护与工厂模式重构                                                                                           │ │                                                                                                                                                            │ │   核心改进:                                                                                                                                                │ │   - 替换过度mock的autouse fixture为精准的线程防护机制                                                                                                      │ │   - 只mock基础设施组件(async_track_time_interval, LifeSmartStateManager)                                                                                   │ │   - 完全保留业务逻辑,允许测试验证真实的成功/失败场景                                                                                                       │ │   - 新增工厂函数模式消除重复mock代码                                                                                                                       │ │                                                                                                                                                            │ │   技术细节:                                                                                                                                                │ │   - conftest.py: 移除auto_mock_lifesmart_hub_setup,新增auto_prevent_thread_creation                                                                        │ │   - test_utils.py: 新增create_mock_oapi_client等工厂函数,集中管理mock创建                                                                                  │ │   - test_hub.py: 更新fixture引用,支持真实Hub逻辑测试                                                                                                       │ │   - 为test_hub.py提供特殊处理,跳过线程mock以测试真实定时器逻辑                                                                                             │ │                                                                                                                                                            │ │   测试覆盖:                                                                                                                                                │ │   - ✅ 单元测试: 验证真实业务逻辑和失败场景                                                                                                                 │ │   - ✅ 集成测试: 使用显式fixture完全控制Hub行为                                                                                                             │ │   - ✅ 线程清理: 从根源防止线程残留,而非事后允许                                                                                                            │ │   - ✅ 全量测试: 854个测试全部通过                                                                                                                          │ │                                                                                                                                                            │ │   设计原则:                                                                                                                                                │ │   - 根本解决线程残留问题而非掩盖症状                                                                                                                       │ │   - 区分单元/集成/隔离测试的不同mock需求                                                                                                                   │ │   - 工厂模式统一mock创建,提高代码复用性                                                                                                                    │ │   - 保持测试可靠性,不掩盖真实的业务逻辑问题
